week3/day 1/book_review_api/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.115.12
googletrans==4.0.2
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
passlib==1.7.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
python-jose==3.4.0
requests==2.32.3
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
